requirements.txt
# Voice Chatbot Requirements
# Install with: pip install -r requirements.txt

# GUI
PyQt6>=6.6.0

# Audio
sounddevice>=0.4.6
numpy>=1.24.0

# Speech-to-Text (faster-whisper)
faster-whisper>=0.10.0

# LLM (Ollama)
ollama>=0.1.0

# Text-to-Speech (Kokoro ONNX v1.0) - English & Spanish
kokoro-onnx>=0.4.0  # Official Kokoro TTS library (v1.0+ for multilingual)
onnxruntime-gpu>=1.16.0  # Use onnxruntime-gpu for CUDA support

# Note: Markdown rendering is done natively with regex - no additional library needed
# The MarkdownRenderer class in main.py handles basic markdown formatting
